+++ lib/arch_guard/arch_guard.mli
@@ -1,0 +1,6 @@
+exception Error of string
+
+module Guard_domain : module type of Guard_domain
+
+val render_json : string list -> string
+val render_text : string list -> string